Rename daemon package to rootd
