+++ pkg/client/rootd/dns/server_windows.go
